django_backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
djangorestframework-simplejwt>=5.2
qrcode>=7.4.2
Pillow>=10.0.0
reportlab>=4.0.0
openpyxl>=3.1.0
pandas>=2.0.0
sendgrid>=6.10.0
python-dotenv>=1.0.0